Exercise C++ from book C++ Basics
Basically  is a calculator on console which works in If_Else way and
Switch way.  Were created to tree modules calculator.cpp(were is
functions, calculator.h(were is the prototypes) and main05.cpp( were the
functions is called).

The exercise is not finished yet. There is one problem which I am going
to fix, problems like the application not continue in switch way.